Refactor page range helpers to eliminate circular dependency
